==> CODE/2. ALS_CF.ipynb <==
import os
import sys

# os.environ['PYSPARK_PYTHON'] = sys.executable
# os.environ['PYSPARK_DRIVER_PYTHON'] = sys.executable

import findspark
findspark.init()

# import libraries
from pyspark import SparkContext
from pyspark.conf import SparkConf
from pyspark.sql import SparkSession
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats
from datetime import datetime
from pyspark.sql import Row
import pyspark.sql.functions as f

import pyspark.sql.types as t

import pyspark.ml.feature as feat

import numpy as np
import pandas as pd

# Pipeline
from pyspark.ml import Pipeline, PipelineModel

# Machine learning
from pyspark.ml.recommendation import ALS
from pyspark.ml.evaluation import RegressionEvaluator

# About Dataset
Use the information:
-  product ID, 
-  User ID, 
- Rating

in dataset Shopee: Products_ThoiTrangNam build a recommended sysytem

**Then make recommendations to some specific users**

SparkContext.setSystemProperty(key='spark.hadoop.dfs.client.use.datanode.hostname',value='true')

sc =SparkContext()

sc.setLogLevel("ERROR")

spark = SparkSession(sc)

df = spark.read.csv(path='../DATA/clean_details.csv',inferSchema=True, header=True)

#sc.cancelAllJobs()

# Overview

df.show(10)
df.printSchema()
print(f'There are {df.count()} rows and {len(df.columns)} columns in dataframe')

## check null

df.select([f.count(f.when(f.isnan(c), c)).alias(c) for c in df.columns]).show()

df.select([f.count(f.when(f.isnull(c), c)).alias(c) for c in df.columns]).show()

## Descirble data

cat_col = df.columns
cat_col

### Category data

#### Check distinct values

df.select([f.count_distinct(c).alias(c) for c in df.columns]).show()

#### Check top5 disticnt value count of each category columns in df 

for col in cat_col:
    (df.groupby(col).count()
     .withColumn('Normalize',(f.col('count')/df.count()))
     .orderBy(f.col('count').desc())
     .show(5,vertical=False,truncate=False))

df = df.withColumnRenamed(existing='user_id',new='usr_id')

- Because my target to recommend the product with high quality, that why reason I only choose rating from 3 to 5 (DID in EDA2)

df.show(5)

#### Drop some product with low rating and drop user with one time buy

##### PRD

group_prd = df.groupby('product_id').agg(f.mean('rating').alias('avg_rating'), f.count('usr_id').alias('count_user'))
group_prd.show(5)

group_prd_gt_3 = group_prd.filter(f.col('avg_rating')>=3)
group_prd_gt_3.show(5)

list_prd = group_prd_gt_3.select('product_id').rdd.flatMap(lambda x: x).collect()
len(list_prd)

list_prd[:5]

##### User Group

group_user = df.groupby('usr_id').agg(f.mean('rating').alias('avg_rating'), f.count('product_id').alias('count_prd'))
group_user.show(5)

group_user_gt_5 = group_user.filter(f.col('count_prd')>=2)
print('Total customer:', group_user_gt_5.count())
group_user_gt_5.show(5)

list_user = group_user_gt_5.select('usr_id').rdd.flatMap(lambda x: x).collect()
print("Total user:",len(list_user))
list_user[:5]

# Preprocessing

final_df = df
final_df.show()

final_df.select([f.count(f.when(f.isnan(c), c)).alias(c) for c in final_df.columns]).show()

final_df.select([f.count(f.when(f.isnull(c), c)).alias(c) for c in final_df.columns]).show()

# Split Data

train, test = final_df.randomSplit([0.8,0.2],seed=42)

# Model

_maxIter = 10
_regParam = 0.01
_rank = 20

als = ALS(maxIter=_maxIter,
          regParam=_regParam, 
          userCol='usr_id',
          itemCol='product_id', 
          ratingCol='rating', 
          coldStartStrategy='drop',
          rank=_rank,
          nonnegative=True)
model_t = als.fit(train)

## Evaluation

# Evaluate the model by computing the RMSE on the test data
predictions = model_t.transform(test)

#inspect
predictions.show(5)

#calcuate rmse score
evaluator = RegressionEvaluator(metricName="rmse",
                                labelCol="rating",
                                predictionCol="prediction")
rmse = evaluator.evaluate(predictions)

print("Root-mean-square error = " + str(rmse))

The model can predict new overall with error nearly 1.0 

## Build Pipeline

pipeline = Pipeline().setStages([als])
p_als_model = pipeline.fit(train)

# save
# models_path =' ../DATA/Pyspark/'
# p_als_model.save(models_path + "/als_model")

# load
# p_als_model = PipelineModel.load(models_path + "/xgb_model.model"

#save model to disk
# model_t.save("../DATA/Pyspark/recommendation_model_amazon_toys_games")

## Providing Recommendations: for all users

# get 5 recommendations which have highest rating.
user_recs = model_t.recommendForAllUsers(5)

user_recs.show(5, truncate=False)

user_recs.select('usr_id').count()

result = user_recs.select(f.col('usr_id'),
                       f.explode(f.col('recommendations')))

result = (result.withColumn('product_id', 
                               result.col.getField('product_id'))
             .withColumn('rating',  
                         result.col.getField('rating'))
            )
result.show()

result_pd = result.select(['usr_id','product_id','rating']).toPandas()
result_pd.head()

for row in result_pd.itertuples(index=False):
    print(row.product_id)
    break

user_id = 31
#create a dictionary for user recommendation
fil_1 = result_pd['usr_id'] == user_id
find_user_rec = result_pd[fil_1]
recom = []
for row in find_user_rec.itertuples(index=False):
    # Access columns using row.column_name
    dict_prd = {'product_id':row.product_id, 'rating':row.rating}
    recom.append(dict_prd)
    #print(row.usr_id, row.product_id, row.rating)

recom_user_dict = {'user_id':user_id,'recom': recom}
recom_user_dict

## Save data to file

#final_df.show()

# Save to disk
#models_path ='../DATA/Pyspark/'
#user_recs.write.parquet('Top_5_User_CF.parquet', mode='overwrite')
#df_asin_asin_idx.write.parquet('reviews_Toys_and_Games_5_Product.parquet', mode='overwrite')

# Define Function for Recommendation

def make_recommendation(usr_id,result_: pd.DataFrame):
    #create a dictionary for user recommendation
    try:
        #create a dictionary for user recommendation
        fil_1 = result_['usr_id'] == user_id
        find_user_rec = result_[fil_1]
        recom = []
        for row in find_user_rec.itertuples(index=False):
            # Access columns using row.column_name
            dict_prd = {'product_id':row.product_id, 'rating':row.rating}
            recom.append(dict_prd)
            #print(row.usr_id, row.product_id, row.rating)
        
        recom_user_dict = {'user_id':user_id,'recom': recom}
        
        return recom_user_dict
    except Exception as e:
        no_value = 'There is no recommendation for this user'
        print(no_value)
        return None

test_dict = make_recommendation(usr_id=31,result_=result_pd)
test_dict

# test = pd.DataFrame(data=test_dict)
# # Expand the column containing lists
# expanded = test['recom'].apply(pd.Series)
# user_recom_df = pd.concat([test.drop(columns=['recom']), expanded], axis=1)
# user_recom_df

def make_result_dataframe(dic_user_rec):
    #adjust rating parameter for filter higher rating
    test = pd.DataFrame(data=dic_user_rec)
    # Expand the column containing lists
    expanded = test['recom'].apply(pd.Series)
    user_recom_df = pd.concat([test.drop(columns=['recom']), expanded], axis=1)
    user_recom_df
    return user_recom_df

make_result_dataframe(dic_user_rec=test_dict)

# Then make recommendations to some users: 

## Get random user

# Get list user ID
full_detail_df = pd.read_csv('../DATA/clean_details.csv')
user_id_unique = full_detail_df['user_id'].unique()

# Get sample user id: 10 values
sample_user = np.random.choice(user_id_unique, size=10, replace=False)
print("Random sample:", sample_user)

sample_user[0]

#make_recommendation(usr_id=102823,result=result)

#final_df.select(f.col("usr_id").alias("random_urs_id")).sample(False, 0.05, seed=0).collect()

#recommendation for each users
for user in sample_user:
    user_id = int(user)
    user_recom = make_recommendation(usr_id=user_id,result_=result_pd)
    if isinstance(user_recom,dict):
        result = make_result_dataframe(dic_user_rec=user_recom)
        print("Recommendation for UserID: ", user_id)
        display(result)

## Save Data Recommedation:

result_pd.to_csv('../DATA/Pyspark/rcm_user_ALS.csv',index=False)
