==> CODE/3. Colaborative_Fitering_Surprise.ipynb <==
import pandas as pd
import numpy as np
# from surprise import Reader, Dataset, SVD, SVDpp, NMF, SlopeOne, KNNBasic, KNNBaseline, KNNWithMeans, KNNWithZScore, CoClustering, BaselineOnly
#from surprise import *
from surprise.model_selection.validation import cross_validate

from surprise import Dataset, Reader, KNNBasic, SVD
from surprise.model_selection import train_test_split

df = pd.read_csv("../DATA/clean_details.csv")

df.info()

#df['product_id'] = df['product_id'].astype('int8')
#df['user_id'] = df['user_id'].astype('int8')
#df['rating'] = df['rating'].astype('float16')

df.head()

df.groupby().describe()

n_ratings = len(df)
n_movies = len(df['product_id'].unique())
n_users = len(df['user_id'].unique())

display(n_ratings, n_movies, n_users)

df['user_id'].value_counts()

df['product_id'].value_counts()

# https://surprise.readthedocs.io/en/stable/reader.html
reader = Reader()
data = Dataset.load_from_df(df[['user_id', 'product_id', 'rating']], reader)

# Define similarity measure
sim_options = {'name': 'cosine', 'user_based': True}

# Split dataset into train and test sets
trainset, testset = train_test_split(data, test_size=0.2)

# Build and train model
#model = KNNBasic(sim_options=sim_options)
model = SVD()
model.fit(trainset)

# Make recommendations for a specific user (replace 'user_id' with the desired user ID)
user_id = 199
# Get items the user hasn't rated yet
items_to_predict = [item_id for item_id in data.build_full_trainset().all_items() if item_id not in trainset.ur[trainset.to_inner_uid(user_id)]]
# Predict ratings for the items
predictions = [model.predict(user_id, item_id) for item_id in items_to_predict]
# Sort predictions by estimated rating
top_n = sorted(predictions, key=lambda x: x.est, reverse=True)[:10]

# Print top recommendations
for prediction in top_n:
    print('Item ID:', prediction.iid, 'Estimated rating:', prediction.est)


# https://www.youtube.com/watch?v=8wLKuscyO9I
# Singular value decomposition
algorithm = SVD()
#algorithm = KNNBasic()
#Run 5-fold cross-validation and print results
results = cross_validate(algorithm, data, measures=['RMSE', 'MAE'], cv=5, verbose=True)

results

# If the results are OK => getting full dataset => fit model
trainset = data.build_full_trainset()
algorithm.fit(trainset)

### Recommender for a specific user

# userId = 27 has seen some movies:
userId = 27
df_select = df[(df['user_id'] == userId) & (df['rating'] >=3)]
df_select = df_select.set_index('product_id')
#df_select = df_select.join(df_title)['Name']
df_select.head(df_select.shape[0])

df_select.shape

df_score = df[["product_id"]]

df_score

- Once the model has been evaluated to our satisfaction, then we can re-train the model using the entire training dataset

# The following are the top 5 movies to be recommended to the user with userId
# To recommend products (i.e., movies) to the given user,
# we can sort the list of movies in decreasing order of predicted ratings
# and take the top N movies as recommendations:
df_score['EstimateScore'] = df_score['product_id'].apply(lambda x: algorithm.predict(userId, x).est) # est: get EstimateScore
df_score = df_score.sort_values(by=['EstimateScore'], ascending=False)

df_score = df_score.drop_duplicates()

df_score.head()

df_score[df_score.EstimateScore>=3]

!
