==> CODE/6. Content_base_Filtering.ipynb <==
%load_ext autoreload
%autoreload 2
from project_transformer import Data_Wrangling

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel, cosine_similarity
from underthesea import word_tokenize, pos_tag, sent_tokenize
import warnings
from gensim import corpora, models, similarities
import gensim
import re

import matplotlib.pyplot as plt 
import seaborn as sns
import glob
import os 
from tqdm.auto import tqdm
import re
import pickle

import scipy.sparse as ss

# Function:

# EMOJI
with open('../DATA/files/emojicon.txt', 'r', encoding="utf8") as file:
    emoji_lst = file.read().split('\n')
    emoji_dict = {}
    for line in emoji_lst:
        key, value = line.split('\t')
        emoji_dict[key] = str(value)

# TEEN CODE
with open('../DATA/files/teencode.txt', 'r', encoding="utf8") as file:
    teen_lst = file.read().split('\n')
    teen_dict = {}
    for line in teen_lst:
        key, value = line.split('\t')
        teen_dict[key] = str(value)

# ENG VIET
with open('../DATA/files/english-vnmese.txt', 'r', encoding="utf8") as file:
    eng_lst = file.read().split('\n')
    eng_dict = {}
    for line in eng_lst:
        key, value = line.split('\t')
        eng_dict[key] = str(value)

# WRONG WORD
with open('../DATA/files/wrong-word.txt', 'r', encoding="utf8") as file:
    wrong_lst = file.read().split('\n')

# STOP WORD
with open('../DATA/files/vietnamese-stopwords.txt', 'r', encoding="utf8") as file:
    stop_lst = file.read().split('\n')

## Build CLEANSER

data_cleanser = Data_Wrangling(emoji_dict=emoji_dict,
                               teen_dict=teen_dict,
                               wrong_lst=wrong_lst,
                               eng_vn_dict=eng_dict,
                               stop_words=stop_lst)

### Gensim

def recommender_gensim(view_prd: str,
                       cleaner,
                       dict_prd: corpora.Dictionary,
                       tfidf_model: gensim.models.tfidfmodel.TfidfModel,
                       similarity: gensim.similarities.docsim.SparseMatrixSimilarity,
                       prd_df: pd.DataFrame):
    # Convert search words into Sparse Vectors
    clean_txt = cleaner.process_text(text=view_prd)
    clean_txt = cleaner.process_postag_thesea(clean_txt)

    clean_ls = clean_txt.split()
    kw_vector = dict_prd.doc2bow(clean_ls)
    print("View product's vector:")
    print(kw_vector)

    # similarity calculation
    sim = similarity[tfidf_model[kw_vector]]

    # print result
    list_id = []
    list_score = []
    for i in range(len(sim)):
        list_id.append(i)
        list_score.append(sim[i])

    # Create df
    df_res = pd.DataFrame({'item_id': list_id, 'score': list_score})

    # Get 5 highest score
    five_h_score = df_res.sort_values(by='score', ascending=False).head(6)
    print("Five highest scores:")
    print(five_h_score)
    print('IDS to list:')
    id_to_ls = five_h_score['item_id'].to_list()
    print(id_to_ls)

    # Find prd
    prd_find = prd_df[prd_df.index.isin(id_to_ls)]
    prd_results = prd_find[['product_id', 'product_name', 'clean_prd_name', 'sub_category', 'price', 'rating']]
    final = pd.concat([prd_results, five_h_score], axis=1).sort_values(by='score', ascending=False)
    final = final[final['score']!=1]
    return final

### Similarity:

def recommend_cosine(text_,
                     df,
                     cleaner,
                     tf_idf_model,
                     tf_idf_arr,
                     num=5):
    import warnings
    warnings.filterwarnings("ignore")
    # Convert search words into Sparse Vectors
    clean_txt = cleaner.process_text(text=text_)
    clean_txt = cleaner.process_postag_thesea(clean_txt)

    # Get Sparse Matrix from TFIDF
    txt_sparse_matrix = tf_idf_model.transform([clean_txt])

    # Calcualte the cosine score
    cosine_text_ = cosine_similarity(X=txt_sparse_matrix,
                                     Y=tf_idf_arr)

    top5_prd = cosine_text_.flatten().argsort()[-num-1:-1]
    top5_score = [(cosine_text_.flatten()[i]) for i in top5_prd]

    # Extract Product
    ls_item = []
    for id, sim in zip(top5_prd, top5_score):
        sim_item = {'item_id': id, 'score': sim}
        ls_item.append(sim_item)
    print(f"San Pham user nhap: {text_}\n\n")
    for rec in ls_item:
        # print(rec[1])
        print(
            f"Recommnended:\tItem ID: {rec['item_id']}, {df.loc[df['item_id'] == rec['item_id'], 'product_name'].values[0]}, (score: {rec['score']})\n")

    # Create DataFrame
    df_res = pd.DataFrame(ls_item)
    filter_ls = df_res['item_id'].to_list()
    prd_find = df[df['item_id'].isin(filter_ls)]
    prd_results = pd.merge(left=prd_find,
                           right=df_res,
                           on=['item_id'],
                           how='inner')
    
    final = prd_results.sort_values(by='score', ascending=False)
    #final = final[final['score']!=1]
    feat = ['product_id', 'product_name', 'clean_prd_name', 'sub_category', 'price', 'rating','item_id','score']
    return final[feat]

def combine_both(gensim_df, cosim_df):
    combine_df = pd.merge(cosim_df,
                     gensim_df,
                     on=['product_id','product_name','clean_prd_name','sub_category','price','rating'],
                     suffixes=('_gensim', '_simi'), 
                     how='outer').drop(columns=['item_id_gensim','item_id_simi'])
    combine_df[combine_df['rating']>3]
    return combine_df

# Load Product DF:

use_cols = ['item_id','product_id','sub_category','price','rating','clean_desc','product_name','clean_prd_name']
clean_prd_df = pd.read_csv('../DATA/final_details.csv',usecols=use_cols)
clean_prd_df.head()

# Gensim

## Load Model and Product Dataframe:

# Load the TF-IDF model from the file
loaded_tfidf_model = models.TfidfModel.load('../DATA/Gensim/tfidf_model')

# Load the dictionary from the file
loaded_dictionary = corpora.Dictionary.load('../DATA/Gensim/corpus_dictionary')

# Load the similarity matrx 
loaded_similarity_matrix  = similarities.SparseMatrixSimilarity.load("../DATA/Gensim/similarity_matrix.index")

## Test Fuction:

#clean_prd_df

text_prd = clean_prd_df.loc[100,'product_name']
text_prd = 'kinh nam rất xịn'
text_prd

results_df = recommender_gensim(view_prd=text_prd,
                                cleaner=data_cleanser,
                                dict_prd=loaded_dictionary,
                                tfidf_model=loaded_tfidf_model,
                                similarity=loaded_similarity_matrix,
                                prd_df=clean_prd_df)

results_df

# Similarity

## Load Model and Product Dataframe:

with open('../DATA/tfidf_vectorizer.pkl', 'rb') as f:
    sim_tfidf_model = pickle.load(f)
type(sim_tfidf_model)

sim_tfidf_matrix = ss.load_npz('../DATA/tfidf_matrix.npz')
type(sim_tfidf_matrix)

## Test Fuction:

#clean_prd_df

#text_prd = clean_prd_df.loc[100,'product_name']
text_prd

sim_res_df = recommend_cosine(text_=text_prd,
                              cleaner=data_cleanser,
                             df=clean_prd_df,
                             tf_idf_model=sim_tfidf_model,
                             tf_idf_arr=sim_tfidf_matrix)

sim_res_df

results_df

text_prd

# Combine BOTH

combine_both(gensim_df=results_df,cosim_df=sim_res_df)

# combine_df = pd.merge(sim_res_df,
#                      results_df,
#                      on=['product_id','product_name','clean_prd_name','sub_category','price','rating'],
#                      suffixes=('_gensim', '_simi'), 
#                      how='outer').drop(columns=['item_id_gensim','item_id_simi'])
# combine_df[combine_df['rating']>3]
